requirements.txt
torch
torchvision
selenium
webdriver-manager
Pillow
requests
tqdm
